src/kontrol/{foundry,display,__main__}: separate out display related module
