+++ src/tests/integration/test_tracing.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from kontrol.foundry import foundry_show
+from kontrol.display import foundry_show
 from kontrol.options import ProveOptions, ShowOptions, TraceOptions
 from kontrol.prove import foundry_prove
